No longer generate a pid file by default (if option not given)
git-svn-id: svn://svn.code.sf.net/p/xymon/code@7925 44351d6e-118b-4698-b696-ce33095ecaa4

+++ branches/4.x-master/xymond/xymond.c
@@ -5973,8 +5973,8 @@ int main(int argc, char *argv[])
 		setsid();
 	}
 
-	/* Save PID */
-	{
+	if (pidfn != NULL) {
+		/* Save PID */
 		FILE *fd = fopen(pidfn, "w");
 		if (fd) {
 			if (fprintf(fd, "%lu\n", (unsigned long)getpid()) <= 0) {
@@ -6313,7 +6313,7 @@ int main(int argc, char *argv[])
 	if (bf_buf) xfree(bf_buf);
 
 	save_checkpoint();
-	unlink(pidfn);
+	if (pidfn) unlink(pidfn);
 
 #ifdef DEBUG_FOR_VALGRIND
 	initialize_hostlist();
